+++ functions/doPush.js
@@ -88,16 +88,16 @@ async function doPush(COMMIT_MESSAGE) {
 	
 	/* BEDROCK ******************************************************************************************************************/
 
-	if (!isEmptyDir(`./texturesPush/Compliance-Bedrock-32x/textures`)) {
-		await autoPush('Compliance-Resource-Pack', 'Compliance-Bedrock-32x', BRANCH_BEDROCK, COMMIT_MESSAGE, './texturesPush/Compliance-Bedrock-32x/')
-		fs.rmdirSync(`./texturesPush/Compliance-Bedrock-32x/textures/`, { recursive: true });
-		console.log(`PUSHED TO GITHUB: Compliance-Bedrock-32x`);
+	if (!isEmptyDir(`./texturesPush/Compliance-Bedrock-32x/1.16.200/textures`)) {
+		await autoPush('Compliance-Resource-Pack', 'Compliance-Bedrock-32x', 'Jappa-1.16.200', COMMIT_MESSAGE, './texturesPush/Compliance-Bedrock-32x/')
+		fs.rmdirSync(`./texturesPush/Compliance-Bedrock-32x/1.16.200/textures/`, { recursive: true });
+		console.log(`PUSHED TO GITHUB: Compliance-Bedrock-32x (1.16.200)`);
 	}
 
-	if (!isEmptyDir(`./texturesPush/Compliance-Bedrock-64x/textures`)) {
-		await autoPush('Compliance-Resource-Pack', 'Compliance-Bedrock-64x', BRANCH_BEDROCK, COMMIT_MESSAGE, './texturesPush/Compliance-Bedrock-64x/')
-		fs.rmdirSync(`./texturesPush/Compliance-Bedrock-64x/textures/`, { recursive: true });
-		console.log(`PUSHED TO GITHUB: Compliance-Bedrock-64x`);
+	if (!isEmptyDir(`./texturesPush/Compliance-Bedrock-64x/1.16.200/textures`)) {
+		await autoPush('Compliance-Resource-Pack', 'Compliance-Bedrock-64x', 'Jappa-1.16.200', COMMIT_MESSAGE, './texturesPush/Compliance-Bedrock-64x/')
+		fs.rmdirSync(`./texturesPush/Compliance-Bedrock-64x/1.16.200/textures/`, { recursive: true });
+		console.log(`PUSHED TO GITHUB: Compliance-Bedrock-64x (1.16.200)`);
 	}
 
 	/* JSON ********************************************************************************************************************/